Enum nat_traversal::HolePunchServerAddr

pub enum HolePunchServerAddr {
    Simple(SocketAddrV4),
    IgdGateway(Gateway),
}

The address of a server that can be used to obtain an external address.

Variants

Simple

A server which speaks the simple hole punching protocol (ie. our MaidSafe protocol). This should probably be deprecated and replaced with a proper STUN implementation.

IgdGateway

An Internet Gateway Device that can be used for UPnP port mapping.