Function nat_traversal::filter_udp_hole_punch_packet

pub fn filter_udp_hole_punch_packet(data: &[u8]) -> Option<&[u8]>

Returns None if data looks like a hole punching message. Otherwise returns the data it was given.

Punching a hole with a udp socket involves packets being sent and received on the socket. After hole punching succeeds it's possible that more hole punching packets sent by the remote peer may yet arrive on the socket. This function can be used to filter out those packets.