Struct nat_traversal::MappedTcpSocket

pub struct MappedTcpSocket {
    pub socket: TcpBuilder,
    pub endpoints: Vec<MappedSocketAddr>,
}

A tcp socket for which we know our external endpoints.

Fields

socket

A bound, but neither listening or connected tcp socket. The socket is bound to be reuseable (ie. SO_REUSEADDR is set as is SO_REUSEPORT on unix).

endpoints

The known endpoints of this socket.

Methods

impl MappedTcpSocket

fn map(socket: TcpBuilder, mc: &MappingContext) -> MappedTcpSocket

Map an existing tcp socket. The socket must not bound or connected. This function will set the options to make the socket address reuseable before binding it.

fn new(mc: &MappingContext) -> MappedTcpSocket

Create a new MappedTcpSocket